ASM-8219 fix pxe_cleanup hanging issue
This issue does not occur all the times, but pxe_cleanup resource
sometimes causes the puppet-agent process to fail because the
ifdown command hangs and fails eventually. This commit modifies
the target resource to call the ifdown command only if the PXE
interface is up and running. Otherwise, it will be ignored.